Rename to clarify that it's client-only
