+++ packages/react-flight/src/ReactFlightClientHostConfigBrowser.js
